+++ parser.d.ts
@@ -126,11 +126,7 @@ type PostprocessEach<I> = I extends `${infer Tag}.${infer Rest}`
   ? PostprocessEach<R>
   : I
 
-/**
- * Internal type.
- * Use at your own risk, since we don't guarantee its signature and stability.
- */
-export type ParseSelectorToTagNames<I extends string> = Trim<I> extends infer I
+type ParseSelectorToTagNames<I extends string> = Trim<I> extends infer I
   ? I extends ''
     ? unknown
     : Split<
@@ -161,3 +157,77 @@ export type TagNameToElement<
   : Tag extends keyof SVGElementTagNameMap
   ? SVGElementTagNameMap[Tag]
   : Fallback
+
+// --------------------------------------------------------
+// Strict Parser
+// --------------------------------------------------------
+
+// Specification is here: https://drafts.csswg.org/css-syntax-3/#ident-token-diagram
+// but we don't plan to comply that fully,
+// otherwise it will increase type-checking time and the complexity of parser.
+
+type LowerCaseLetter =
+  | 'a'
+  | 'b'
+  | 'c'
+  | 'd'
+  | 'e'
+  | 'f'
+  | 'g'
+  | 'h'
+  | 'i'
+  | 'j'
+  | 'k'
+  | 'l'
+  | 'm'
+  | 'n'
+  | 'o'
+  | 'p'
+  | 'q'
+  | 'r'
+  | 's'
+  | 't'
+  | 'u'
+  | 'v'
+  | 'w'
+  | 'x'
+  | 'y'
+  | 'z'
+
+type Digit = '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9'
+
+type IdentifierFirstChar =
+  | LowerCaseLetter
+  | Uppercase<LowerCaseLetter>
+  | '-'
+  | '_'
+type IdentifierChar = IdentifierFirstChar | Digit
+
+type IsValidTags<S> = S extends [infer H, ...infer Rest]
+  ? H extends '' | '*'
+    ? IsValidTags<Rest>
+    : IsValidTagName<H> extends true
+    ? IsValidTags<Rest>
+    : false
+  : true
+type IsValidTagName<S> = S extends `${infer H}${infer Rest}`
+  ? H extends IdentifierFirstChar
+    ? IsValidRestChars<Rest>
+    : false
+  : false
+type IsValidRestChars<S extends string> = S extends `${infer H}${infer Rest}`
+  ? H extends IdentifierChar
+    ? IsValidRestChars<Rest>
+    : false
+  : true // no characters left, so it's OK
+
+export type StrictlyParseSelector<S extends string> =
+  ParseSelectorToTagNames<S> extends infer Tags
+    ? Tags extends []
+      ? TagNameToElement<''>
+      : Tags extends string[]
+      ? IsValidTags<Tags> extends true
+        ? TagNameToElement<Tags[number]>
+        : never
+      : never
+    : never