+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v2.8.0
+
 - Export "strict" parser. (Close [#25](https://github.com/g-plane/typed-query-selector/issues/25))
 - Removed exported private type `ParseSelectorToTagNames`.